+++ R/tools.R
@@ -68,11 +68,12 @@ makeBlankModel <- function(modelInputs) {
     covariance       = NA,
     numObs           = NA,
     numParams        = NA,
+    call             = NA,
+    freq             = NA,
     iterations       = NA,
     message          = NA,
     standardDraws    = NA,
-    randParSummary   = NA,
-    status           = -1,
+    status           = -99,
     modelType        = modelInputs$modelType,
     modelSpace       = modelInputs$modelSpace,
     priceName        = modelInputs$priceName,
@@ -80,9 +81,13 @@ makeBlankModel <- function(modelInputs) {
     randPars         = modelInputs$randPars,
     parSetup         = modelInputs$parSetup,
     weightsUsed      = modelInputs$weightsUsed,
-    options          = modelInputs$options
+    clusterName      = modelInputs$clusterName,
+    numClusters      = modelInputs$numClusters,
+    robust           = modelInputs$robust,
+    standardDraws    = NA,
+    options          = options
   ),
-  class = c("logitr", "logitr.fail")
+  class = "logitr"
   )
   return(result)
 }
@@ -144,32 +149,9 @@ ci <- function(data, alpha = 0.025) {
 is_logitr <- function(x) {
   inherits(x, "logitr")
 }
-is_logitr_multistart <- function(x) {
-  inherits(x, "logitr.multistart")
-}
-is_logitr_allRuns <- function(x) {
-  inherits(x, "logitr.allRuns")
-}
-is_logitr_fail <- function(x) {
-  inherits(x, "logitr.fail")
-}
 
-allRunsCheck <- function(model) {
-  if (is_logitr_allRuns(model)) {
-    model <- useBestModel(model)
-  }
-  return(model)
-}
-
-# Return the best model, and print a warning statement
-useBestModel <- function(model) {
-  model <- model$bestModel
-  message(
-    "NOTE: Using results from run ", model$multistartNumber, " of ",
-    model$options$numMultiStarts,
-    " multistart runs\n(the run with the largest log-likelihood value)"
-  )
-  return(model)
+isMxlModel <- function(parSetup) {
+  return(("n" %in% parSetup) | ("ln" %in% parSetup))
 }
 
 # Functions for getting specific parameter indexes
@@ -187,8 +169,4 @@ getNormParIDs <- function(parSetup) {
 
 getLogNormParIDs <- function(parSetup) {
   return(which(parSetup == "ln"))
-}
-
-isMxlModel <- function(parSetup) {
-  return(("n" %in% parSetup) | ("ln" %in% parSetup))
 }