+++ R/wtp.R
@@ -19,39 +19,21 @@
 #' @return A data frame of the WTP estimates.
 #' @export
 #' @examples
-#' # Run a MNL model in the Preference Space:
 #' library(logitr)
 #'
+#' # Estimate a preference space model
 #' mnl_pref <- logitr(
-#'   data = yogurt,
+#'   data       = yogurt,
 #'   choiceName = "choice",
-#'   obsIDName = "obsID",
-#'   parNames = c("price", "feat", "dannon", "hiland", "yoplait")
+#'   obsIDName  = "obsID",
+#'   parNames   = c("price", "feat", "brand")
 #' )
 #'
-#' # Get the WTP implied from the preference space model
+#' # Compute the WTP implied from the preference space model
 #' wtp(mnl_pref, priceName = "price")
 wtp <- function(model, priceName) {
-  if (is_logitr(model) == FALSE) {
-    stop('Model must be estimated using the"logitr" package')
-  }
-  if (is.null(priceName)) {
-    stop("Must provide priceName to compute WTP")
-  }
-  model <- allRunsCheck(model)
-  if (model$modelSpace == "pref") {
-    return(getPrefSpaceWtp(model, priceName))
-  } else if (model$modelSpace == "wtp") {
-    wtp_mean <- stats::coef(model)
-    wtp_se <- model$standErrs
-    return(getCoefSummaryTable(
-      wtp_mean, wtp_se, model$numObs, model$numParams))
-  }
-}
-
-getPrefSpaceWtp <- function(model, priceName) {
-  # Compute mean WTP
-  coefs <- stats::coef(model)
+  wtpInputsCheck(model, priceName)
+  coefs <- model$coef
   priceID <- which(names(coefs) == priceName)
   pricePar <- -1 * coefs[priceID]
   wtp_mean <- coefs / pricePar
@@ -63,7 +45,7 @@ getPrefSpaceWtp <- function(model, priceName) {
   wtpDraws <- draws / priceDraws
   wtpDraws[, priceID] <- draws[, priceID]
   wtp_se <- apply(wtpDraws, 2, stats::sd)
-  return(getCoefSummaryTable(wtp_mean, wtp_se, model$numObs, model$numParams))
+  return(getCoefTable(wtp_mean, wtp_se))
 }
 
 #' Compare WTP from preference and WTP space models
@@ -88,41 +70,38 @@ getPrefSpaceWtp <- function(model, priceName) {
 #' WTP space models.
 #' @export
 #' @examples
-#' # Run a MNL model in the Preference Space:
 #' library(logitr)
 #'
+#' # Estimate a MNL model in the Preference space
 #' mnl_pref <- logitr(
-#'   data = yogurt,
+#'   data       = yogurt,
 #'   choiceName = "choice",
-#'   obsIDName = "obsID",
-#'   parNames = c("price", "feat", "dannon", "hiland", "yoplait")
+#'   obsIDName  = "obsID",
+#'   parNames   = c("price", "feat", "brand")
 #' )
 #'
-#' # Get the WTP implied from the preference space model
+#' # Compute the WTP implied from the preference space model
 #' wtp_mnl_pref <- wtp(mnl_pref, priceName = "price")
 #'
-#' # Run a MNL model in the WTP Space:
+#' # Estimate a MNL model in the WTP Space, using the computed WTP values
+#' # from the preference space model as starting points
 #' mnl_wtp <- logitr(
-#'   data = yogurt,
+#'   data       = yogurt,
 #'   choiceName = "choice",
-#'   obsIDName = "obsID",
-#'   parNames = c("feat", "dannon", "hiland", "yoplait"),
+#'   obsIDName  = "obsID",
+#'   parNames   = c("feat", "brand"),
 #'   priceName = "price",
 #'   modelSpace = "wtp",
 #'   options = list(startVals = wtp_mnl_pref$Estimate)
 #' )
 #'
-#' # Compare the WTP between the two spaces:
+#' # Compare the WTP between the two spaces
 #' wtpCompare(mnl_pref, mnl_wtp, priceName = "price")
 wtpCompare <- function(model_pref, model_wtp, priceName) {
-  if (is_logitr(model_pref) == FALSE | is_logitr(model_wtp) == FALSE) {
-    stop('Models must be estimated using the "logitr" package')
-  }
-  model_pref <- allRunsCheck(model_pref)
-  model_wtp <- allRunsCheck(model_wtp)
+  wtpCompareInputsCheck(model_pref, model_wtp, priceName)
   pref <- wtp(model_pref, priceName)$Estimate
   pref <- c(pref, model_pref$logLik)
-  wtp <- stats::coef(model_wtp)
+  wtp <- model_wtp$coef
   wtp <- c(wtp, model_wtp$logLik)
   names(pref)[length(pref)] <- "logLik"
   names(wtp)[length(wtp)] <- "logLik"